parser/requirements.txt
telethon>=1.36
aiosqlite>=0.20
python-dotenv>=1.0
